[UI] Code refactoration
